api/requirements.txt
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
docopt==0.6.2
donttrust==0.1.5
Flask==2.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pipreqs==0.4.13
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.5
Werkzeug==2.3.7
yarg==0.1.9
Flask-Cors==4.0.0
smartsheet-python-sdk==3.0.2